refactor: updated project structure and gave it a fresh start
