--- backend/.env
@@ -1,3 +1,0 @@
-PORT=5000
-MONGODB_URI=mongodb://localhost:27017/event-notification
-JWT_SECRET=your_jwt_secret_key_here
