--- backend/Dockerfile
@@ -1,23 +1,0 @@
-# Use official Node.js LTS image
-FROM node:18-alpine
-
-# Create app directory
-WORKDIR /usr/src/app
-
-# Copy package files
-COPY package*.json ./
-
-# Install dependencies
-RUN npm install
-
-# If you're in development and want to use nodemon
-# RUN npm install --only=development
-
-# Copy app source
-COPY . .
-
-# Expose the port the app runs on
-EXPOSE 5000
-
-# Command to run the application
-CMD ["npm", "start"]
