--- backend/controllers/notificationController.js
@@ -1,168 +1,0 @@
-const Notification = require('../models/Notification');
-const { StatusCodes } = require('http-status-codes');
-
-// @desc    Get all notifications for a user
-// @route   GET /api/notifications
-// @access  Private
-const getNotifications = async (req, res) => {
-  try {
-    const { userId } = req.user;
-    const { limit = 10, page = 1 } = req.query;
-    const skip = (page - 1) * limit;
-
-    const notifications = await Notification.find({ recipient: userId })
-      .sort({ createdAt: -1 })
-      .skip(skip)
-      .limit(parseInt(limit));
-
-    const total = await Notification.countDocuments({ recipient: userId });
-
-    res.status(StatusCodes.OK).json({
-      success: true,
-      count: notifications.length,
-      total,
-      pages: Math.ceil(total / limit),
-      data: notifications,
-    });
-  } catch (error) {
-    console.error('Error getting notifications:', error);
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-      success: false,
-      message: 'Server error',
-    });
-  }
-};
-
-// @desc    Mark notification as read
-// @route   PATCH /api/notifications/:id/read
-// @access  Private
-const markAsRead = async (req, res) => {
-  try {
-    const { id } = req.params;
-    const { userId } = req.user;
-
-    const notification = await Notification.findOneAndUpdate(
-      { _id: id, recipient: userId },
-      { isRead: true },
-      { new: true, runValidators: true }
-    );
-
-    if (!notification) {
-      return res.status(StatusCodes.NOT_FOUND).json({
-        success: false,
-        message: 'Notification not found',
-      });
-    }
-
-    res.status(StatusCodes.OK).json({
-      success: true,
-      data: notification,
-    });
-  } catch (error) {
-    console.error('Error marking notification as read:', error);
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-      success: false,
-      message: 'Server error',
-    });
-  }
-};
-
-// @desc    Mark all notifications as read
-// @route   PATCH /api/notifications/read-all
-// @access  Private
-const markAllAsRead = async (req, res) => {
-  try {
-    const { userId } = req.user;
-
-    await Notification.updateMany(
-      { recipient: userId, isRead: false },
-      { isRead: true }
-    );
-
-    res.status(StatusCodes.OK).json({
-      success: true,
-      message: 'All notifications marked as read',
-    });
-  } catch (error) {
-    console.error('Error marking all notifications as read:', error);
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-      success: false,
-      message: 'Server error',
-    });
-  }
-};
-
-// @desc    Delete a notification
-// @route   DELETE /api/notifications/:id
-// @access  Private
-const deleteNotification = async (req, res) => {
-  try {
-    const { id } = req.params;
-    const { userId } = req.user;
-
-    const notification = await Notification.findOneAndDelete({
-      _id: id,
-      recipient: userId,
-    });
-
-    if (!notification) {
-      return res.status(StatusCodes.NOT_FOUND).json({
-        success: false,
-        message: 'Notification not found',
-      });
-    }
-
-    res.status(StatusCodes.OK).json({
-      success: true,
-      data: {},
-    });
-  } catch (error) {
-    console.error('Error deleting notification:', error);
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-      success: false,
-      message: 'Server error',
-    });
-  }
-};
-
-// @desc    Create a new notification (for internal use)
-// @route   POST /api/notifications
-// @access  Private/Admin
-const createNotification = async (req, res) => {
-  try {
-    const { recipient, title, message, type, event, metadata } = req.body;
-
-    const notification = new Notification({
-      recipient,
-      title,
-      message,
-      type: type || 'other',
-      event,
-      metadata: metadata || {},
-    });
-
-    await notification.save();
-
-    // Here you would typically emit a WebSocket event to notify the client in real-time
-    // io.to(`user_${recipient}`).emit('new_notification', notification);
-
-    res.status(StatusCodes.CREATED).json({
-      success: true,
-      data: notification,
-    });
-  } catch (error) {
-    console.error('Error creating notification:', error);
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-      success: false,
-      message: 'Server error',
-    });
-  }
-};
-
-module.exports = {
-  getNotifications,
-  markAsRead,
-  markAllAsRead,
-  deleteNotification,
-  createNotification,
-};
