--- backend/middleware/auth.js
@@ -1,53 +1,0 @@
-const jwt = require('jsonwebtoken');
-const { StatusCodes } = require('http-status-codes');
-const User = require('../models/User');
-
-// Protect routes
-const protect = async (req, res, next) => {
-  let token;
-
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith('Bearer')
-  ) {
-    try {
-      // Get token from header
-      token = req.headers.authorization.split(' ')[1];
-
-      // Verify token
-      const decoded = jwt.verify(token, process.env.JWT_SECRET);
-
-      // Get user from the token
-      req.user = await User.findById(decoded.id).select('-password');
-      next();
-    } catch (error) {
-      console.error('Error verifying token:', error);
-      res.status(StatusCodes.UNAUTHORIZED).json({
-        success: false,
-        message: 'Not authorized, token failed',
-      });
-    }
-  }
-
-  if (!token) {
-    res.status(StatusCodes.UNAUTHORIZED).json({
-      success: false,
-      message: 'Not authorized, no token',
-    });
-  }
-};
-
-// Grant access to specific roles
-const authorize = (...roles) => {
-  return (req, res, next) => {
-    if (!roles.includes(req.user.role)) {
-      return res.status(StatusCodes.FORBIDDEN).json({
-        success: false,
-        message: `User role ${req.user.role} is not authorized to access this route`,
-      });
-    }
-    next();
-  };
-};
-
-module.exports = { protect, authorize };
