--- backend/models/Notification.js
@@ -1,47 +1,0 @@
-const mongoose = require('mongoose');
-
-const notificationSchema = new mongoose.Schema(
-  {
-    recipient: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'User',
-      required: true,
-    },
-    title: {
-      type: String,
-      required: true,
-      trim: true,
-    },
-    message: {
-      type: String,
-      required: true,
-    },
-    type: {
-      type: String,
-      enum: ['event_reminder', 'event_update', 'system', 'other'],
-      default: 'other',
-    },
-    isRead: {
-      type: Boolean,
-      default: false,
-    },
-    event: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: 'Event',
-    },
-    metadata: {
-      type: Object,
-      default: {},
-    },
-  },
-  {
-    timestamps: true,
-  }
-);
-
-// Index for faster querying
-notificationSchema.index({ recipient: 1, isRead: 1, createdAt: -1 });
-
-const Notification = mongoose.model('Notification', notificationSchema);
-
-module.exports = Notification;
