+++ backend/package.json
@@ -1,22 +1,29 @@
 {
-  "name": "event-notification-backend",
+  "name": "pec-events-server",
   "version": "1.0.0",
-  "description": "Backend for Event Notification System",
-  "main": "server.js",
+  "description": "",
+  "main": "index.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
+    "start": "node dist/index.js",
+    "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+    "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "commonjs",
   "dependencies": {
-    "express": "^4.18.2",
-    "mongoose": "^7.0.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express-validator": "^7.0.0",
-    "http-status-codes": "^2.2.0"
+    "dotenv": "^17.2.1",
+    "express": "^5.1.0",
+    "mongoose": "^8.16.5"
   },
   "devDependencies": {
-    "nodemon": "^2.0.20"
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.3",
+    "@types/node": "^24.1.0",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.8.3"
   }
 }
