+++ docker-compose.yml
@@ -1,85 +1,65 @@
-version: '3.8'
+# version: '3.8'
 
 services:
-  # MongoDB Service
-  mongodb:
-    image: mongo:6.0
-    container_name: event-notification-db
-    environment:
-      - MONGO_INITDB_DATABASE=event-notification
-    volumes:
-      - mongodb_data:/data/db
-    ports:
-      - "27017:27017"
-    networks:
-      - app-network
-
-  # Backend Service
-  backend:
-    build:
-      context: ./backend
-      dockerfile: Dockerfile
-    container_name: event-notification-backend
-    environment:
-      - NODE_ENV=development
-      - MONGODB_URI=mongodb://mongodb:27017/event-notification
-      - JWT_SECRET=your_jwt_secret_key_here
-    ports:
-      - "5000:5000"
-    depends_on:
-      - mongodb
-    networks:
-      - app-network
-    volumes:
-      - ./backend:/usr/src/app
-      - /usr/src/app/node_modules
-    restart: unless-stopped
-
-  # Frontend Service
-  frontend:
-    build:
+  app:
+    build: 
       context: ./frontend
-      dockerfile: Dockerfile
-    container_name: event-notification-frontend
-    environment:
-      - NODE_ENV=development
-      - REACT_APP_API_URL=http://localhost:5000/api
+      dockerfile: Dockerfile.dev
     ports:
-      - "19000:19000"  # Expo
+      - "8081:8081"
+      - "19000:19000"  # Expo dev server
       - "19001:19001"  # Expo dev tools
-      - "19002:19002"  # Expo dev server
-      - "19006:19006"  # Expo Web
+      - "19002:19002"  # Expo dev tools
+      - "19006:19006"  # Metro bundler
     volumes:
       - ./frontend:/app
       - /app/node_modules
-    depends_on:
-      - backend
-    networks:
-      - app-network
-    stdin_open: true
-    tty: true
-    restart: unless-stopped
+      - /app/.expo
+    environment:
+      - NODE_ENV=development
+      - EXPO_DEVTOOLS_LISTEN_ADDRESS=0.0.0.0
+      - EXPO_DEV_SERVER_ORIGIN=http://localhost:19000
+      - EXPO_TUNNEL_SUBDOMAIN=event-notification
+      - EXPO_DEBUG=false
+      - EXPO_NO_WARNINGS=1
+      - EXPO_NO_TTY=1
+    stdin_open: true  # Keep STDIN open for interactive mode
+    tty: true         # Allocate a pseudo-TTY
+    # depends_on:
+    #   - backend
 
-  # Nginx Reverse Proxy (optional)
-  nginx:
-    image: nginx:alpine
-    container_name: event-notification-proxy
-    ports:
-      - "80:80"
-    volumes:
-      - ./nginx/nginx.conf:/etc/nginx/nginx.conf:ro
-    depends_on:
-      - backend
-      - frontend
-    networks:
-      - app-network
-    restart: unless-stopped
+#   backend:
+#     build: 
+#       context: ./backend
+#       dockerfile: Dockerfile.dev
+#     ports:
+#       - "5000:5000"
+#     volumes:
+#       - ./backend:/app
+#       - /app/node_modules
+#     environment:
+#       - NODE_ENV=development
+#       - MONGODB_URI=mongodb://mongodb:27017/eventdb
+#     depends_on:
+#       - mongodb
 
-# Define networks
-networks:
-  app-network:
-    driver: bridge
+#   mongodb:
+#     image: mongo:6.0
+#     ports:
+#       - "27017:27017"
+#     volumes:
+#       - mongodb_data:/data/db
+#     restart: always
 
-# Define volumes
-volumes:
-  mongodb_data:
+  # nginx:
+  #   image: nginx:alpine
+  #   ports:
+  #     - "80:80"
+  #   volumes:
+  #     - ./nginx/nginx.conf:/etc/nginx/nginx.conf
+  #   depends_on:
+  #     - frontend
+#       - backend
+
+# volumes:
+#   mongodb_data:
