--- frontend/App.tsx
@@ -1,195 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { Ionicons } from '@expo/vector-icons';
-import * as Notifications from 'expo-notifications';
-import { Platform } from 'react-native';
-import { StatusBar } from 'expo-status-bar';
-
-// Screens
-import LoginScreen from './src/screens/LoginScreen';
-import RegisterScreen from './src/screens/RegisterScreen';
-import NotificationListScreen from './src/screens/NotificationListScreen';
-import NotificationDetailScreen from './src/screens/NotificationDetailScreen';
-import SettingsScreen from './src/screens/SettingsScreen';
-import { AuthProvider, useAuth } from './src/context/AuthContext';
-
-// Types
-type RootStackParamList = {
-  Login: undefined;
-  Register: undefined;
-  Main: undefined;
-  NotificationDetail: { notificationId: string };
-};
-
-type MainTabParamList = {
-  Notifications: undefined;
-  Settings: undefined;
-};
-
-const Stack = createNativeStackNavigator<RootStackParamList>();
-const Tab = createBottomTabNavigator<MainTabParamList>();
-
-// Configure notifications
-Notifications.setNotificationHandler({
-  handleNotification: async () => ({
-    shouldShowAlert: true,
-    shouldPlaySound: true,
-    shouldSetBadge: true,
-  }),
-});
-
-// Main Tab Navigator
-function MainTabs() {
-  return (
-    <Tab.Navigator
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ focused, color, size }) => {
-          let iconName: keyof typeof Ionicons.glyphMap = 'notifications';
-
-          if (route.name === 'Notifications') {
-            iconName = focused ? 'notifications' : 'notifications-outline';
-          } else if (route.name === 'Settings') {
-            iconName = focused ? 'settings' : 'settings-outline';
-          }
-
-          return <Ionicons name={iconName} size={size} color={color} />;
-        },
-        tabBarActiveTintColor: '#007AFF',
-        tabBarInactiveTintColor: 'gray',
-      })}
-    >
-      <Tab.Screen 
-        name="Notifications" 
-        component={NotificationListScreen} 
-        options={{ title: 'Notifications' }}
-      />
-      <Tab.Screen 
-        name="Settings" 
-        component={SettingsScreen} 
-        options={{ title: 'Settings' }}
-      />
-    </Tab.Navigator>
-  );
-}
-
-// Main App Component
-function AppContent() {
-  const { user, isLoading } = useAuth();
-  const [expoPushToken, setExpoPushToken] = useState('');
-  const notificationListener = useRef<Notifications.Subscription>();
-  const responseListener = useRef<Notifications.Subscription>();
-
-  // Request notifications permission and get push token
-  useEffect(() => {
-    registerForPushNotificationsAsync().then(token => {
-      if (token) {
-        setExpoPushToken(token);
-        // Here you would typically send this token to your backend
-        // to associate it with the user's device
-      }
-    });
-
-    // Handle notifications received while the app is in the foreground
-    notificationListener.current = Notifications.addNotificationReceivedListener(notification => {
-      // Handle the notification
-      console.log('Notification received:', notification);
-    });
-
-    // Handle user tapping on a notification
-    responseListener.current = Notifications.addNotificationResponseReceivedListener(response => {
-      console.log('Notification response received:', response);
-      // You could navigate to a specific screen based on the notification
-    });
-
-    return () => {
-      if (notificationListener.current) {
-        Notifications.removeNotificationSubscription(notificationListener.current);
-      }
-      if (responseListener.current) {
-        Notifications.removeNotificationSubscription(responseListener.current);
-      }
-    };
-  }, []);
-
-  if (isLoading) {
-    return null; // Or a loading spinner
-  }
-
-  return (
-    <NavigationContainer>
-      <StatusBar style="auto" />
-      <Stack.Navigator>
-        {user ? (
-          <>
-            <Stack.Screen 
-              name="Main" 
-              component={MainTabs} 
-              options={{ headerShown: false }}
-            />
-            <Stack.Screen 
-              name="NotificationDetail" 
-              component={NotificationDetailScreen}
-              options={{ title: 'Notification Details' }}
-            />
-          </>
-        ) : (
-          <>
-            <Stack.Screen 
-              name="Login" 
-              component={LoginScreen} 
-              options={{ headerShown: false }}
-            />
-            <Stack.Screen 
-              name="Register" 
-              component={RegisterScreen}
-              options={{ title: 'Create Account' }}
-            />
-          </>
-        )}
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
-}
-
-// Main App Wrapper with Auth Provider
-export default function App() {
-  return (
-    <AuthProvider>
-      <AppContent />
-    </AuthProvider>
-  );
-}
-
-// Helper function to register for push notifications
-async function registerForPushNotificationsAsync() {
-  let token;
-  
-  if (Platform.OS === 'android') {
-    await Notifications.setNotificationChannelAsync('default', {
-      name: 'default',
-      importance: Notifications.AndroidImportance.MAX,
-      vibrationPattern: [0, 250, 250, 250],
-      lightColor: '#FF231F7C',
-    });
-  }
-
-  const { status: existingStatus } = await Notifications.getPermissionsAsync();
-  let finalStatus = existingStatus;
-  
-  if (existingStatus !== 'granted') {
-    const { status } = await Notifications.requestPermissionsAsync();
-    finalStatus = status;
-  }
-  
-  if (finalStatus !== 'granted') {
-    console.log('Failed to get push token for push notification!');
-    return null;
-  }
-  
-  token = (await Notifications.getExpoPushTokenAsync()).data;
-  console.log('Expo push token:', token);
-  
-  return token;
-}
