+++ frontend/Dockerfile.dev
@@ -1,0 +1,30 @@
+FROM node:18-alpine
+
+WORKDIR /app
+
+# Install Expo CLI and ngrok globally
+RUN npm install -g expo-cli @expo/ngrok@^4.1.0
+
+# Set environment variables for Expo
+ENV EXPO_DEVTOOLS_LISTEN_ADDRESS=0.0.0.0
+ENV EXPO_DEV_SERVER_ORIGIN=http://localhost:19000
+ENV EXPO_USE_STATIC_WEB_PACKAGER=true
+ENV EXPO_TUNNEL_SUBDOMAIN=event-notification
+ENV EXPO_DEBUG=false
+ENV EXPO_NO_WARNINGS=1
+ENV EXPO_NO_TTY=1
+
+# Copy package files
+COPY package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy source code
+COPY . .
+
+# Expose ports
+EXPOSE 19000 19001 19002 8081
+
+# Start development server with tunnel
+CMD ["npx", "expo", "start", "--tunnel"]
