+++ frontend/app.json
@@ -1,38 +1,34 @@
 {
   "expo": {
-    "name": "Event Notification App",
-    "slug": "event-notification-app",
+    "scheme": "peceventapp",
+    "name": "pec-event-app",
+    "slug": "pec-event-app",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "light",
+    "newArchEnabled": true,
     "splash": {
-      "image": "./assets/splash.png",
+      "image": "./assets/splash-icon.png",
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
-    "updates": {
-      "fallbackToCacheTimeout": 0
-    },
-    "assetBundlePatterns": [
-      "**/*"
-    ],
     "ios": {
-      "supportsTablet": true,
-      "bundleIdentifier": "com.yourevent.notificationapp"
+      "supportsTablet": true
     },
     "android": {
       "adaptiveIcon": {
         "foregroundImage": "./assets/adaptive-icon.png",
-        "backgroundColor": "#FFFFFF"
+        "backgroundColor": "#ffffff"
       },
-      "package": "com.yourevent.notificationapp"
+      "edgeToEdgeEnabled": true
     },
     "web": {
+      "bundler": "metro",
       "favicon": "./assets/favicon.png"
     },
-    "extra": {
-      "apiUrl": "http://localhost:5000/api"
-    }
+    "plugins": [
+      "expo-router"
+    ]
   }
 }
