+++ nginx/nginx.conf
@@ -3,39 +3,33 @@ events {
 }
 
 http {
-    upstream backend {
-        server backend:5000;
+    upstream frontend {
+        server frontend:8081;
     }
 
-    upstream frontend {
-        server frontend:19000;
-    }
+    # upstream backend {
+    #     server backend:5000;
+    # }
 
     server {
         listen 80;
         server_name localhost;
 
-        # Backend API
-        location /api/ {
-            proxy_pass http://backend:5000/;
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;
-            proxy_set_header Connection 'upgrade';
+        location / {
+            proxy_pass http://frontend;
             proxy_set_header Host $host;
-            proxy_cache_bypass $http_upgrade;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             proxy_set_header X-Forwarded-Proto $scheme;
         }
 
-        # Frontend
-        location / {
-            proxy_pass http://frontend:19000;
-            proxy_http_version 1.1;
-            proxy_set_header Upgrade $http_upgrade;
-            proxy_set_header Connection 'upgrade';
-            proxy_set_header Host $host;
-            proxy_cache_bypass $http_upgrade;
-        }
+        # location /api {
+        #     rewrite ^/api/(.*) /$1 break;
+        #     proxy_pass http://backend;
+        #     proxy_set_header Host $host;
+        #     proxy_set_header X-Real-IP $remote_addr;
+        #     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        #     proxy_set_header X-Forwarded-Proto $scheme;
+        # }
     }
 }
